+++ src/rules/no-unused-package-exports/rule.ts
@@ -1,13 +1,26 @@
+import type { JSONSchema4 } from '@typescript-eslint/utils/json-schema';
+import { z } from 'zod';
+
 import { createRule, getESMInfo, getLocFromRange } from '../util.js';
 
-export const noUnusedPackageExports = createRule({
+const schema = z
+  .strictObject({
+    ignorePackages: z.array(z.string()).optional(),
+  })
+  .optional();
+type Options = z.infer<typeof schema>;
+
+export const noUnusedPackageExports = createRule<
+  [Options],
+  'noUnusedPackageExports'
+>({
   name: 'no-unused-package-exports',
   meta: {
     docs: {
       description:
         'Ensure entry point exports are imported elsewhere in the monorepo',
     },
-    schema: [],
+    schema: [schema.toJSONSchema() as JSONSchema4],
     fixable: undefined,
     type: 'problem',
     messages: {
@@ -15,6 +28,7 @@ export const noUnusedPackageExports = createRule({
         'Export point export "{{name}}" must be imported in another package',
     },
   },
+  defaultOptions: [{ ignorePackages: [] }],
   create(context) {
     const esmInfo = getESMInfo(context);
 
@@ -25,7 +39,7 @@ export const noUnusedPackageExports = createRule({
       return {};
     }
 
-    const { fileInfo } = esmInfo;
+    const { fileInfo, packageInfo } = esmInfo;
     /* istanbul ignore if */
     if (fileInfo.fileType !== 'code') {
       return {};
@@ -34,6 +48,15 @@ export const noUnusedPackageExports = createRule({
     // If this file doesn't have any entry points, then we can bail early and
     // avoid needing to do a more expensive filter.
     if (!fileInfo.entryPointSpecifier) {
+      return {};
+    }
+
+    // Check if this package is ignored, and if so bail
+    const { ignorePackages = [] } = context.options[0] ?? {};
+    if (
+      packageInfo.packageName &&
+      ignorePackages.includes(packageInfo.packageName)
+    ) {
       return {};
     }